+++ decode/plot.py
@@ -23,8 +23,7 @@ from logging import getLogger
 import numpy as np
 import matplotlib.pyplot as plt
 from scipy.signal import hanning
-from .utils.ndarray.functions import psd, allan_variance
-from .utils.misc.functions import deprecation_warning
+from .utils import allan_variance, deprecation_warning, psd
 
 
 # module logger
