--- decode/utils/__init__.py
@@ -1,9 +1,0 @@
-# coding: utf-8
-# flake8: noqa
-
-
-from .misc.functions import *
-from .ndarray.functions import *
-
-del misc
-del ndarray
